for test

+++ api-rabbit/src/main/java/com/yunsoo/api/rabbit/dto/MktDrawPrize.java
@@ -3,9 +3,9 @@ package com.yunsoo.api.rabbit.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.yunsoo.common.data.databind.DateTimeJsonDeserializer;
-import com.yunsoo.common.data.databind.DateTimeJsonSerializer;
 import com.yunsoo.common.data.object.MktDrawPrizeObject;
+import com.yunsoo.common.databind.DateTimeJsonDeserializer;
+import com.yunsoo.common.databind.DateTimeJsonSerializer;
 import org.joda.time.DateTime;
 
 /**
@@ -30,7 +30,7 @@ public class MktDrawPrize {
     private String drawRuleId;
 
     @JsonProperty("amount")
-    private Integer amount;
+    private Double amount;
 
     @JsonProperty("mobile")
     private String mobile;
@@ -103,11 +103,11 @@ public class MktDrawPrize {
         this.drawRuleId = drawRuleId;
     }
 
-    public Integer getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(Integer amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
